+++ Carplate-yolov5/README.md
@@ -11,18 +11,30 @@ Python 3.8 or later with all [requirements.txt](https://github.com/CertifaiAI/ca
 ```bash
 $ git clone https://github.com/CertifaiAI/car-plate-recognition.git
 $ cd Carplate-yolov5
-$ pip install -r requirements.txt
+$ bash install_scripts.sh
 ```
-
-### Download Weights
-
-Due to the Github memory limit, the weights have been uploaded on to [Google Drive](https://drive.google.com/drive/folders/1afPFDv9Fo0GW4W5ss6GWgBGX31iUmn4t), please download the weights and place it inside of the 'weights/' folder in this project directory.
 
 ### Run the Server
 
 Before running the scripts make sure to first setup the FastAPI backend server, to do so please refer to the following [instructions](https://github.com/CertifaiAI/car-plate-recognition/blob/main/Backend-server/README.MD).
 
 ### Inference
+*main.py*: Runs the cargate program as a whole (includes sensors). User can select which sensors to use
+```bash
+$ python main.py --show
+```  
+This show output result on PC.
+
+```bash
+$ python main.py --show --nano
+```  
+This show output result on Jetson Nano.
+
+```bash
+$ python main.py --show --nano --sensor --led --relay
+```  
+This show output result on Jetson Nano.
+
 
 *send_images.py*: Iterates through a number of pre-saved NumberPlate images, sends the images over to the FastAPI backend server and outputs the LPR. 
 
@@ -36,3 +48,8 @@ $ python send_images.py
 $ python send_video.py
 ```
 
+### Run with GPU
+In order to run using GPU, please modify device on Carplate-yolov5/config.py
+```
+self.DEVICE = 'cpu'
+```
