+++ Carplate-yolov5/install_torch.sh
@@ -1,0 +1,1 @@
+pip3 install torch==1.9.0+cpu torchvision==0.10.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
