+++ Carplate-yolov5/ledPanel.py
@@ -1,0 +1,10 @@
+'''
+    Send data to bluetooth device to display data on P10 LED Panel
+'''
+
+class LedPanel():
+    def __init__(self):
+        pass
+
+    def send_data(self, data):
+        pass
