+++ Carplate-yolov5/requirements.txt
@@ -1,25 +1,15 @@
-# pip install -r requirements.txt
-
 # base ----------------------------------------
-matplotlib>=3.2.2
-numpy>=1.18.5
-opencv-python>=4.1.2
-Pillow
-PyYAML>=5.3.1
-scipy>=1.4.1
-torch>=1.7.0
-torchvision>=0.8.1
-tqdm>=4.41.0
-
-# logging -------------------------------------
-tensorboard>=2.4.1
+requests
+opencv-python
 
 # plotting ------------------------------------
-seaborn>=0.11.0
 pandas
-
-# extras --------------------------------------
-thop  # FLOPs computation
 
 # sensors -------------------------------------
 Jetson.GPIO
+
+# Status report -------------------------------
+paho-mqtt
+
+# Install weights from drive
+gdown
